# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API sistem absensi mahasiswa berbasis geolocation. Semua respons memakai envelope {success, data, message, errors}.",
        "title": "Absensi Mahasiswa API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/schedules": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Schedules"
                ],
                "summary": "Daftar jadwal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter kelas",
                        "name": "class_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter mata kuliah",
                        "name": "course_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Schedules"
                ],
                "summary": "Buat jadwal",
                "parameters": [
                    {
                        "description": "Jadwal",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            }
        },
...
        "/schedules/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Schedules"
                ],
                "summary": "Detail jadwal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID jadwal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Schedules"
                ],
                "summary": "Ubah jadwal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID jadwal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Jadwal",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Schedules"
                ],
                "summary": "Hapus jadwal (soft delete)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID jadwal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            }
        },
        "/schedules/{id}/generate-code": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Schedules"
                ],
                "summary": "Generate kode absensi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID jadwal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Masa berlaku (menit, 1-1440)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.GenerateCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            }
        },
        "/schedules/{id}/generate-qr": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Schedules"
                ],
                "summary": "Rotate token QR jadwal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID jadwal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            }
        },
        "/schedules/{id}/qr.png": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "Admin: Schedules"
                ],
                "summary": "Gambar QR PNG jadwal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID jadwal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.GenerateCodeRequest": {
            "type": "object",
            "properties": {
                "minutes_valid": {
                    "type": "integer",
                    "maximum": 1440,
                    "minimum": 1
                }
            }
        },
...
        "dto.ScheduleRequest": {
            "type": "object",
            "required": [
                "class_id",
                "course_id",
                "end_time",
                "location_id",
                "start_time"
            ],
            "properties": {
                "class_id": {
                    "type": "integer"
                },
                "course_id": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "is_recurring": {
                    "type": "boolean"
                },
                "location_id": {
                    "type": "integer"
                },
                "recurrence_end": {
                    "type": "string"
                },
                "repeat_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: \"Bearer {access_token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}